+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - [BUGFIX] Now we can cancel editing key
 - [BUGFIX] Fix reconnect issues
 - [BUGFIX] Fix synchronization start after first login
+- [Feature] Write and emulate buttons view
 - [Feature] Add update Flipper via OTA
 - [Feature] Added an option to cancel existing Flipper RPC requests
 - [Feature] Use new design for unsupported card and for unsupported dialog